[MAC80211][app][Update ated and iwpriv  wrapper for interface conversion]
[Description]
Add interface conversion to the ated wrapper and change the method of detecting the sku of user's board
to reading eeprom in debugfs instead of the lspci command (not included in external release).
Update interface conversion mechanism which considers ATECTRLBANDIDX in iwpriv wrapper.
Add work mode conversion in iwpriv wrapper

[Release-log]
N/A

Change-Id: Ia93731ff8de84b528c4aa273245da5850e830358