[mt7988][cpu][Add cache hierarchy device tree node]
[Description]
Add CPU cache hierarchy node to device tree. This prevents the kernel
from displaying the following warning message during boot-up:

"cacheinfo: Unable to detect cache hierarchy for CPU 0"

This warning message usually causes users to misunderstand that the
CPU cache is broken, when in reality it is caused by bus hanging.

[Release-log]
N/A


Change-Id: I0afbad319f658c8cb8e3a677a29f464abe51ba6c

Change-Id: Ie78266d1816bd994a21b5a2788a398c46463c427