[kernel][common][spi][spi-mt65xx: Fix dual/quad mode for IPM design]
[Description]
Fix the issue of dts buswidth cannot be applied properly in spi-mt65xx.
(1) Fix the name of buswidth to bus-width in dts in order to fit the
format in linux spi kernel so that spi-tx-bus-width & spi-rx-bus-width
can be parsed properly.
(2) Add quad and dual ability of transfer mode in spim controller.
We have known that spi->mode is determined by dts & spi controller
mode->bits. However, the SPI TX/RX DUAL/QUAD bits haven't been set now
in spim controller and this leads to the fact that dts buswidth cannot
be applied properly.

[Release-log]
N/A


Change-Id: I6cf32362bf825a8d1871b3d64c653740df04c7ff