[][kernel][mt7988][Fix USXGMII to SGMII mode change issue]
[Description]
Fix USXGMII to SGMII mode change issue.

In the mt7988, the USXGMII and SGMII share the same PHYA.
In order to make sure the SGMII driver can reconfigure PHYA, it has to
check PCS link status when phylink calling to mtk_sgmii_pcs_config().

If without this patch, the SGMII driver might not be able to reconfigure
PHYA when changing the USXGMII to the SGMII.

[Release-log]


Change-Id: Iab56d70ac32bc0f6b4c7235027ac0ebb430a4a01
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7464574