[][kernel][mt7988][Fix SGMII to USXGMII mode change issue for the 10G SFP+ optical module]
[Description]
Fix SGMII to USXGMII mode change issue for the 10G SFP+ optical module.

Due to the USXGMII hardware limitation, it has to make sure receiving
RX signal before calibration.
Therefore, the USXGMII driver has to reconfigure PHYA until link up with
link partner.

If without this patch, the USXGMII mignt not be able to link up with
10G SFP+ optical module.

[Release-log]
N/A


Change-Id: I76fbde6298daac3610af0b85d1d95ca0b537018e
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7465697