[checkpatch: fix check failures for backport patches]
[Description]
Fix checkpatch check failures for backport patches.

With the release of Linux kernel v5.7, the default character limit was
increased to 100. This change was implemented in following commit:
bdc48fa ("checkpatch/coding-style: deprecate 80-column warning").

To ensure successful addition of backport patches to mtk-openwrt-feeds,
we need to set the $max_line_length variable to 100 instead of 80. This
will prevent checkpatch failures when backporting patches to our SDK.

[Release-log]
N/A


Change-Id: I1889420907d61da3ca34d1fc6966ab169fffd114