[][Kernel][hnat][Fix Veth TX packets issue]
[Description]
Fix Veth TX packets issue.
When veth TX packets, MAC of the packets were incorrectly modified.

HNAT only supports TCP/UDP packets received from WIFI or Ethernet.
The supported packets will be marked in skb hnat headroom.
All hnat hook function should check whether it carries this flag.
If it is not, it should not be processed and should be returned in advance.

[Release-log]
N/A


Change-Id: I08e813d472f3cc345b3ae5267a5f350267fbb569
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7662944