[kernel][common][eth][Add MIB read feature to the Aquantia PHY driver]
[Description]
Add MIB read feature to the Aquantia PHY driver.

The users can utilize following command to dump the MIB information
from AQR113C.
  - cat /sys/kernel/debug/aquantia/phyX; X is 0 or 8

This feature is available only on the link partners with speeds of
2.5G, 5G, and 10G.

Without this patch, it is difficult to locate the source of packet
loss, whether it is on the system side or the line side.

[Release-log]
N/A


Change-Id: I863b3af761993f08a10f770632e2885f50f4a33a