[][MAC80211][hnat][Fix 2.5G LAN/WAN low t-Put issue for the PPPQ]
[Description]
Fix 2.5G LAN/WAN low t-Put issue for the PPPQ.

In the current PPPQ design, the Lan0~5 ports are limited to 1Gbps.
Consequently, the 2.5G Lan5 is also restricted to 1Gbps, leading to
poor DL and Bi-Di performance.
Therefore, we exclude the 2.5G WAN to 2.5G Lan5 case from the PPPQ in
order to avoid an unnecessary rate-limited operation.

Without this patch, the user may experience the poor perforamce in
the 2.5G LAN/WAN scenario.

[Release-log]
N/A


Change-Id: I6178feaab929a1c6dc14ca738a66e94c04afea9d
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7905304