[kernel][common][eth][Update RX ring debug command]
[Description]
Refactor RX ring debug command.

In the previous commit [0], we have added full support for the RSS.
However, we did not update the RX ring debug command, which means that
users are unable to dump the descriptor from all of the RX rings.

[0] https://git01.mediatek.com/plugins/gitiles/openwrt/feeds/
mtk-openwrt-feeds/+/94806ecf62f6e2c511dc5e5be69867d41ea12671

Without this patch, users are unable to dump the descriptor from all of
the RX rings.

[Release-log]
N/A


Change-Id: I3210fc382d0d88a75e44cb88dc08c67f7c397623