[][kernel][mt7986/mt7981][eth][Add Ethernet RSS support for Panther and Cheetah]
[Description]
Add Ethernet RSS (receive side scaling) support for Panther and Cheetah.

The main purpose of RSS is to increase the utilization rate
of each CPU as much as possible to prevent the performance
from not being fully demonstrated.

In order to improve software path peak throughput, it's
necessary to enble RSS to hash Rx packets among multi-cores.

If without this patch, Panther and Cheetah software path can't meet peak
throughput requirement.

[Release-log]
N/A


Change-Id: Ib6b18a8ab6a5580a85a3c68964c2b41906e25cfb
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7986183