[MAC80211][hnat][Fix ETH bring up issue for the mt7622]
[Description]
Fix ETH bring up issue for the mt7622.

The mtk_probe() function cannot request an interrupt because some
legacy platforms do not have individual PDMA interrupts.
Therefore, we add an MTK_PDMA_INT flag to bypass requesting the PDMA
interrupt in this case.

Without this patch, the ETH cannot be brought up on the MT7622.

[Release-log]
N/A


Change-Id: Ib88e7744b004dfadf6738ab20a3d012157d9c34d