[MAC80211][hnat][Improve ETH to ETH performace in an unbalanced PHY rate test for the mt7988]
[Description]
Add two pachtes to improve ETH to ETH performace in an unbalanced PHY
rate test for the mt7988.

These patches will enlarge FQ DMA size to 4K and dispatch short packets
a high priority queue in the PPPQ path.

Without this patch, the performance of ETH to ETH cannot reach line rate
in an unbalanced PHY rate test for the mt7988.

[Release-log]
N/A


Change-Id: I572c886af40fbce236ba93ccfa1118b246f608d7