[kernel][common][eth][Fix refcount underflow issue in the pending_work]
[Description]
Fix refcount underflow issue in the pending_work.

If one of the Ethernet interface is down, the pending_work should not
call to mtk_stop() function again to avoid a refcount underflow.

Without this patch, the ETH may experience the kernel panic which
caused by a refcount underflow.

[Release-log]
N/A


Change-Id: I4e1f40640c09372cddad57994da03f70107f133a