[MAC80211][WiFi6][hostapd][Fix mu_onoff was overwritten with unexpected values]
[Description]
Fix vendor comamnd configuration "mu_onoff" was overwritten with driver default values.

After interface is up, the mu_onoff within the hostapd configuration are
read by hostapd. If the AP is running on DFS channel, the setting will
only be sent to the driver by hostapd after the CAC is completed.
If $ hostapd_cli get_mu is used to read the mu_onoff from the
driver before CAC is completed, the "mu_onoff" value replied by the
driver will be written the the current setting of hostapd. This case
will result in the user-configured "mu_onoff" value being overwritten by
the driver's default value.

With this commit, user can use get_mu command only when state is HAPD_IFACE_ENABLED.

[Release-log]
N/A

Change-Id: I989304fd66f7e8cf805d352d1575449e53bdf2ef