[][kernel][mt7988][eth][Fix Link Partner link down issue for the SFP+ optical module]
[Description]
Fix Link Partner link down issue for the SFP+ optical module.

In certain scenarios, users may frequently check the status of the
link. This operation invokes the phylink_resolve() function and can
potentially race with the phylink's link_poll thread. As a result,
it is necessary to introduce a spinlock to safeguard the accuracy
of the USXGMII configuration sequence.

Without this patch, the Link Partner might disconnect during the
reboot test.

[Release-log]
N/A


Change-Id: I46e4091e2da6f05a87725ea2229e40703a5a5d03
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8381615