[][openwrt][mt7988][crypto][Refactor FP setting in IPSec hnat binding flow]
[Description]
Refactor FP setting in IPSec hnat binding flow.
In the previous version, FP was determined based on the destination
device during binding. For instance, if the destination device is a WAN,
then FP is directly set to the GMAC2 port.
However, in mt7988d configuration, there may be a swapping of LAN port
and WAN port. This means that WAN can either be GMAC2 or GMAC3.
Therefore, we now rely on the interface id to determine which FP should
be set.

[Release-log]
N/A


Change-Id: I2bda377a9c668b02f29a9fcd3425b4e65d22a26a
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8383420