[MAC80211][hnat][Update IPv6 unbind issue for the routing mode]
[Description]
Refactor IPv6 unbind issue for the routing mode.

In the previous commit [0], we fixed an issue where flow offload
could not be bound in IPv6 routing mode. But this commit caused
the IPv4 cannot bind in ETH to WiFi bridge mode.

Therefore, we refactor the previous commit to resolve this issue.

[0] https://git01.mediatek.com/plugins/gitiles/openwrt/feeds/
mtk-openwrt-feeds/+/494fbb6115ed8fa6b2f68296b35337c92690d2ee

Without this patch, the IPv6 flow cannot bind in the routing mode.

[Release-log]
N/A


Change-Id: I61a274952be686c6b0ebc8a30edf3e8cc12df94e