[kernel][mt7988][eth][Fix panic issue during SER test on the 4GB RFB]
[Description]
Fix panic issue during SER test on the 4GB RFB.

If the mtk_poll_rx() function detects the MTK_RESETTING flag, it will
jump to release_desc and refill the high word of the SDP on the 4GB RFB.
Subsequently, mtk_rx_clean will process an incorrect SDP, leading to a
panic.

Without this patch, the kernel will panic in the mtk_rx_clean() function
during SER test.

[Release-log]
N/A


Change-Id: I097ff1b8898dde20edfd7631f626a85582bfc955