[][openwrt][mt7988][crypto][Refactor tport setting for packets to EIP197]
[Description]
Refactor tport setting for packets to EIP197.
In previous version, we use skb_hnat_cdrt and magic tag to determine
whether a packet should be send to EIP197. However, for packets that
performed segement in GSO, headroom data is not consistent with the
original packet. To ensure that only inline mode packets be sent to
EIP197. We use skb->inner_protocol for determination in this patch.
XFRM callback function for inline mode will set the skb->inner_protocol.

[Release-log]
N/A


Change-Id: I6ef1d7f2044ead78d8a756bc012226573758069d
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8494301