[][MAC80211][WiFi6][mt76][Fix bug in VoW DebugFS command]
[Description]
Fix a bug in DebugFS command "vow".
Specifically, variable "rv" is not properly initialized, so its value might be random, causing the conditional statement in "show_vow_info" command to fail.
To fix this bug, do not check the number of arguments for "show_vow_info" command.
In addition, coding style is refined.

[Release-log]
N/A

Change-Id: I346e4eaa27ab82f924c5f10cb1f481a83eddc231
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8547829