[][kernel][common][net][Add AN8855 dsa and gsw driver]
[Description]
Add AN8855 dsa and gsw driver.
AN8855 is a 5-port giga switch with 1 HSGMII port5 as the CPU port.

This patch adds the dsa driver for modern OpenWRT support, and gsw
driver to support legacy swconfig.

phy addr of 5 internal gphy is 1 to 5 in default. User can set the
"changesmiaddr" and "airoha,smi-addr" in the dts file to change the
beginning number of the phy addr. For example,
    set changesmiaddr = <6>;   // dsa driver
    or  airoha,smi-addr = <6>; // gsw driver
    then 5 phy addr will be 6 to 10.

[Release-log]
N/A

Change-Id: I7d14afde33bf556e5a7b223258f8a68de36dd0c2
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8758316