[MAC80211][hnat][Improve ETH to ETH performace in an unbalanced PHY rate test for mt7981/7986]
[Description]
Change patch to improve ETH to ETH performace in an unbalanced PHY
rate test for mt7981/7986

This patch would enable to dispatch the short packet into a higher
priority HQoS queue in PPPQ mode, which used to only enable in MT7988.

Without this patch, the performance of ETH to ETH cannot reach line rate
in an unbalanced PHY rate test for the mt7981/7986.

[Release-log]
N/A


Change-Id: I7fe14d1c2a975767311c3e281478f4933c65ff7d