[CRITICAL][kernel][mt7988][eth][Fix issue of PSE hanging when the internal 2.5G PHY connection is disconnected]
[Description]
Fix issue of PSE hanging when the internal 2.5G PHY connection is
disconnected.

If the internal 2.5G PHY connection is disconnected, it will gate
the clock of XGMAC, causing any remaining packets to stuck in the
XGMAC FIFO and PSE output queue, ultimately leading to the PSE
becoming unresponsive.

Without this patch, the PSE may experiecnce a hanging issue when
the internal 2.5G PHY connection is disconnected.

[Release-log]
N/A


Change-Id: I4bafd9e5b36f5f5028e1a72e6f1681099a923a63