[MAC80211][WiFi6][mt76][Fix inconsistent QoS mapping between SW and HW]
[Description]
Fix inconsistent QoS mapping between SW and HW.
Specifically, the mapping from IP DSCP to IEEE 802.11 user priority may be customized.
Therefore, driver needs to pass the mapping to HW, so that the QoS type of traffic can be mapped in a consistent manner for both SW and HW paths.

[Release-log]
N/A

Change-Id: I3ba8a74b1978c6554747494be30ba25383cad379