[][MAC80211][WiFi6][core][Add exported function for SoftMAC driver to get QoS map]
[Description]
Add exported function for SoftMAC driver to get QoS map.
Because the mapping from IP DSCP to IEEE 802.11 user priority may be customized.
Therefore, driver needs to pass the mapping to HW, so that the QoS type of traffic can be mapped in a consistent manner for both SW and HW paths.

[Release-log]
N/A

Change-Id: I3a37b0c5d109751be58bb57c99665d6b92992b3a
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8999115