[HIGH][kernel][common][eth][Refactor the link down recovery handler for out-of-band devices used with USXGMII/SGMII]
[Description]
Refactor the link down recovery handler for out-of-band devices used
with USXGMII/SGMII.

In the previous design, the link down handler used to be executed only
after the Line side link was up. However, certain PHY devices like
EN8811H do not release the Line side until the System side link is up.
To address this issue, we have introduced methods pcs_enable() and
pcs_disable() to carry out a link pool procedure. This procedure will
monitor the system-side link status every second for out-of-band devices.

Without this patch, the SGMII might experience a link down issue when
connected to the EN8811H.

[Release-log]
N/A


Change-Id: I3f07b7fa206a5d3dcfad0fc9c3f05251c04a0803