[kernel][mt7988][eth][Refactor the support for the permanent MAC address of the ethtool]
[Description]
Refactor the support for the permanent MAC address of the ethtool.

In the previous design, we allowed the kernel to change the driver
loading sequence if Flash is not ready to read perm_addr, which could
potentially lead to some driver dependency issues. As a result, we
relocated the reading of perm_addr to the mtk_open() function.

Without this patch, the kernel may encounter a driver dependency issue
when enabling the perm_addr reading feature.

[Release-log]
N/A


Change-Id: I7eaa8f02d326ca7c98d6d740b894d0aea5a5fc85