[kernel][mt7981][eth][Refactor issue of GMAC RX hanging when PHYs operates in half-duplex mode]
[Description]
Fix issue of GMAC RX hanging when PHYs operates in half-duplex mode.

In the previous update, we enabled PRMBL_LMT to filter out irregular
short packets when the internal GPHY PHY is connected to the legacy
10Mbps hub. However, we are facing a similar issue with the SGMII
PHY(EN8801S). As a result, we will also enable PRMBL_LMT function
in this scenario.

Without this patch, the GMAC RX direction might be stalled due to
irregular short packets.

[Release-log]
N/A


Change-Id: I2490e2945cafcdddc918ad217aabddbf4d191107