[kernel][mt7988][hnat][Change PPE entry IPV6_HNAPT condition to support NPTv6]
[Description]
Change PPE entry IPV6_HNAPT condition to support NPTv6.

The previous solution to determine if the entry is IPv6 routing or
IPv6 HNAPT was based on nf_conntrack status, but NPTv6 is not supported
by nf_conntrack. Therefore, we changed the condition to check if the
ingress IPv6 address and the egress IPv6 address are the same.

[Release-log]
N/A


Change-Id: Ia02249120b38fac0e74a79b58ac115ff6bf0ca06