[HIGH][kernel][common][eth][Fix incorrect TX DMA buffer length]
[Description]
Fix incorrect TX DMA buffer length.

The DMA buffer length is used to calculate the number of
TX DMA descriptors needed for a single TX. If the size is
incorrect, it will lead to incorrect calculation results.

Without this patch, there is a higher probability of the
software queue being stopped.

[Release-log]
N/A


Change-Id: Ideb77c7412c0c7df6b001d3a4d4aebd37a2ea585