[HIGH][kernel][common][hnat][Fix issue nf_conntrack statistics update when dump all_entry]
[Description]
Fix issue nf_conntrack statistics update when dump all_entry

In @9116932, we store the nfct of skb for further nf_conntrack statistics
update when the entry is set to BIND, but it would be a problem when
the memory of the conntrack have been garbage collected.

In this patch we try to use tuple(ip address, tcp/udp port) to get the
conntrack and update the mib counter statistics to nf conntrack to avoid
the problem in above.

[Release-log]
N/A


Change-Id: I4707b32c60faed4553c6c58eee131b02b1d53507