[kernel][mt7988][eth][Update 36-bit DMA feature to resolve panic issue]
[Description]
Refactor 36-bit DMA feature to resolve panic issue.

In the previous change with ID @9236901, we refactor the 36-bit DMA
feature to an upstream style, inadvertently introducing a potential
risk of losing the upper 4 bits of the DMA addressin the mtk_poll_rx()
function during SER.

Without this patch, the kernel will panic in the mtk_rx_clean() function
during SER test.

[Release-log]
N/A


Change-Id: Id35a034f27239289fa9380e5bc4dfddd1e3f2d81