[HIGH][kernel][common][eth][Change the network device TX queue reset flow]
[Description]
Change the network device TX queue reset flow.

The purpose of resetting the TX queue is to reset the
byte and packet count as well as to clear the software
flow control XOFF bit.

Although the original reset flow is upstream code, we
found that it only resets queue 0 of each network device.
Queues that are not reset may cause unexpected issues.

Without this patch, packets may stop being sent after
Netsys reset and "transmit timeout" log may be displayed.

[Release-log]
N/A


Change-Id: Id968ef0cc3fa91339e8b65e3951a803f079347cf