[kernel][common][eth][Refactor phylink pcs_enable and pcs_disable methods]
[Description]
Refactor phylink pcs_enable and pcs_disable methods.

In the scenario where 'pcs' is not present, the temporary memory 'pcs'
might remain uninitialized. However, if pl->pcs is not a NULL pointer,
it will cause the CPU to unexpectedly execute pcs_enable, leading to
the kernel dereferencing a non-NULL pointer.

Without this patch, the kernel might encounter a panic issue when
calling the pcs_enable function.

[Release-log]
N/A


Change-Id: I454c1e5cf15ce5ac00f08c73abfae4f1d9e47f47