[][kernel][mt7988][eth][Refactor QDMA rate limit for the TOPS]
[Description]
Refactor QDMA rate limit for the TOPS.

Due to the previous method, some of phylink's functions for getting
the maximum speed for rate adaptation were removed. Therefore, it
causes the GMAC to be unable to configure the speed correctly when
the PHY is running in rate adaptation mode.

Without this path, the GMAC is unable to configure the speed correctly
when the PHY is running in rate adaptation mode.

[Release-log]
N/A


Change-Id: If380e6f1bc9f80eaa3ba89c3ebb13dc854b13984
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/9632660