[kernel][common][eth][Change the wait time for SER handshake with WiFi timeout to 10 seconds, and refactor the debugging messages]
[Description]
Change the wait time for SER handshake with WiFi timeout to 10 seconds,
and refactor the debugging messages.

In current design, the NETSYS SER only waits for 3 seconds for WiFi to
stop traffic. Therefore, NETSYS SER might encounter unexpected results
if users print a lot of messages in the WiFi SER handler.

[Release-log]
N/A


Change-Id: I9363591be005a7db419c4cf70977626a44cafbd2