[kernel][common][app][Refactor the mtk_ppe support for the mtkhnat_util]
[Description]
Refactor the mtk_ppe support for the mtkhnat_util.

If the autobuild does not enable mt_wifi or mtk_warp, the mtkhnat
module will not be loaded. As a result, mtkhnat_utils will misjudge
the HNAT module, causing mtkhnat_util to be unable to set the HQoS
setting for the QDMA.

[Release-log]
N/A


Change-Id: I2d2ffa1a2d2c56a545aa0b1c8fc91e1b81f3cc39