[kernel][mt7988][eth][Refactor the MTK_HWLRO enabling method by relocating it to the MT7988_CAPS]
[Description]
Refactor the MTK_HWLRO enabling method by relocating it to the MT7988_CAPS.

The mtk_hw_init() function uses MT7988_CPAS to determine whether the
Dummy Page needs to be enabled or not. Therefore, the current method
may mistakenly cause the driver to enable the Dummy Page feature.

Without this patch, the driver may mistakenly enable the Dummy Page
feature.

[Release-log]
N/A


Change-Id: I0995e06aec878ef74e7d816d5fd53a7d66c64854