[kernel][common][eth][Fix traffic stuck issue with EN8811H after the network restart]
[Description]
Fix traffic stuck issue with EN8811H after the network restart.

This is a legacy API that switches to the 1000BaseX interface when the
PHY link speed is not at 2500Mbps during a network restart command on
the EN8811H.

Without this patch, the SGMII connection will go down with the EN8811H
when the PHY link speed is not at 2500Mbps.

[Release-log]
N/A


Change-Id: I42aa1afdca43f5cbae48d7cd6b8b64715d42bf51