[kernel][mt7987][hnat][Refactor fqos flag assignment when HQoS is enabled for the NETSYSv3.1]
[Description]
Refactor fqos flag assignment when HQoS is enabled for the NETSYSv3.1.

In MT7987, tport is retired, to enable HQoS function it would
depend on fqos. Since in Netsys 3.0(MT7988&MT7987) before this patch,
only the tport would be set to 1 when HQoS is enabled.

In this patch, it would align the fqos flag setting condition
to tport_id in order to enable HQoS in MT7987.

[Release-log]
N/A


Change-Id: I5fef2fe751361961ee87fd4a057d95270f78c759