[kernel][mt7987][eth][Add validation for SRAM allocation and move FQ_DMA's DMAD to SRAM]
[Description]
Add validation for SRAM allocation and move FQ_DMA's DMAD to SRAM.

Because mt7987 has only 128KB of SRAM, the ETH driver cannot
allocate all the TX, RX, and FQ DMA's DMAD to the SRAM. Hence, we
have implemented a validation for the SRAM allocation to verify if
the SRAM address is valid or not. If it is not valid, the driver
will fall back to DRAM allocation.

Without this patch, the ETH driver cannot enable the has_sram feature
on the mt7987.

[Release-log]
N/A


Change-Id: I52923b15b7a446547db70881eea79b5d3741437d