[kernel][common][eth][Refactor NETSYS SER to prevent resetting GMAC when GMAC is not hang]
[Description]
Refactor NETSYS SER to prevent resetting GMAC when GMAC is not hang.

In the previous design, NETSYS SER would reset the FE and GMAC in any
hang condition. However, in cases where the GMAC is not hung, we can
use MAC_LINK_DOWN instead of a PHY disconnect to save a few seconds of
waiting for the PHY link to come up, thereby avoiding a WiFi traffic
stop.

Without this patch, the ETH to WiFi traffic would stop for a few seconds
even if the GMAC is not hang condition.

[Release-log]
N/A


Change-Id: I584a5a325a86e587fe5f9d8a558ec12458548a68