[][MAC80211][eth][Refactor the QoS configuration to use the existing shared functions]
[Description]
Refactor the QoS configuration to use the existing shared functions.

In the previous change, we relocated the NF HNAT's QoS relevant
functions to mtk_eth_dbg.c. As a result, it is now possible to share
these functions with Flowblock HNAT.
Therefore, we created symbolic links to mtketh instead of implementing
individual functions in mtk_qdma_debugfs.c.

[Release-log]
N/A


Change-Id: I1c065ae16624c0d47f3928e489231d3bcea24fbf
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/9790085