[HIGH][kernel][mt7988][eth][Refactor the issue of PSE hanging when the internal 2.5G PHY connection is disconnected]
[Description]
Refactor the issue for PSE hanging when the internal 2.5G PHY
connection is disconnected.

If the internal 2.5G PHY connection is disconnected, it will gate
the clock of XGMAC, causing any remaining packets to stuck in the
XGMAC FIFO and PSE output queue, ultimately leading to the PSE
becoming unresponsive.

The formal solution has been added to the internal 2.5G PHY firmware
version 20241101, so we can remove the previous solution now.

[Release-log]
N/A


Change-Id: Ia54f4859141898308f1277c0d31860cf709859c5