[kernel][common][eth][Add the SGMII/USXGMII link_poll disable feature into the ETH debugfs]
[Description]
Add the SGMII/USXGMII link_poll disable feature into the ETH debugfs.

Sometimes we need to measure the SI when bringing up the new board, but
link_poll will try to reconfigure the SGMII/USXGMII mode upon detecting
a link down or a manual change in mode. Therefore, we add a link_poll
disable feature to deactivate this mechanism for SI measurement.
 - Disable link poll for the SGMII0
   echo 0 0 > /sys/kernel/debug/mtketh/sgmii_link_poll;
 - Disable link poll for the SGMII1
   echo 1 0 > /sys/kernel/debug/mtketh/sgmii_link_poll;
 - Disable link poll for the USXGMII0
   echo 0 0 > /sys/kernel/debug/mtketh/usxgmii_link_poll;
 - Disable link poll for the USXGMII1
   echo 1 0 > /sys/kernel/debug/mtketh/usxgmii_link_poll;

[Release-log]
N/A


Change-Id: I87e2335370e6975b285dae617cc48620d281a276