[kernel][common][eth][Refactor error detection rule for the WDMA_RX monitor]
[Description]
Refactor error detection rule for the WDMA_RX monitor.

We have found that the WDMA CRX/DRX will stop for more than 100 seconds
when the DUT is in a very noisy environment. Therefore, we will try to
reduce false alarms by increasing the maximum hang count to 180 seconds
for the CONNSYS_BUSY error condition.

Without this patch, the CONNSYS_BUSY error condition will trigger false
alarms in the WDMA_RX monitor when DUT is in a very noisy environment.

[Release-log]
N/A


Change-Id: I7eaebbc13e84c36818ff84d24935b26738806e47