[kernel][common][eth][Add a new case for the forced trigger of NETSYS SER]
[Description]
Add a new case for the forced trigger of NETSYS SER.

In the current design, we use a 'force' flag to bypass the forced
trigger of SER. However, some situations cannot be bypassed, such
as switching from XGDM to GDM and WiFi driver sending an
MTK_FE_START_RESET_INIT event. Therefore, we have added a new case
to handle these cases.

After this change, we can observe the following behavior.
- force = 0, only the cases mentioned above can trigger SER
- force = 1, DMA monitor and cases mentioned above can trigger SER

[Release-log]
N/A


Change-Id: Iba4241503f9185de0673f19982556bd7e32bf055