[kernel-6.6][mt7988][eth][Add support for the MXL86252 RFB with one Ethernet passive Mux]
[Description]
Add support for the MXL86252 RFB with one Ethernet passive Mux.

In order to support SFP/PHY dynamic switching, we add a passive mux to
the SerDes path that connects with the SFP cage and PHY. We then use
the GPIO to control this mux, which will switch to the SFP framework
when the SFP module is inserted and will switch to the PHY framework
when the SFP module is removed.

Without this patch, the ETH driver is unable to support Ethernet passive
Mux.

[Release-log]
N/A


Change-Id: Ifd2dd7d760ee0b2db764aa98894dd91c0d132d51