[kernel][common][eth][Refactor the power down setting for the SGMII]
[Description]
Refactor the power down setting for the SGMII.

In the hardware design, the SGMII hardware will automatically set Bit3
and Bit0 of offset 0xE8 when the driver sets Bit4 of offset 0xE8.
However, if the driver only sets Bit4, it may cause a race condition
with the SGMII hardware. Therefore, the designer suggests setting Bit4,
Bit3, and Bit0 simultaneously.

[Release-log]
N/A


Change-Id: Ic2bf9b34f0f70b97197557d84e5fda33555915bd