[kernel][common][eth][Refactor the register dump for the NETSYS warm reset]
[Description]
Refactor the register dump for the NETSYS warm reset.

In the current flow, we did not dump the necessary registers after
preparing for a reset. Therefore, we have changed the location of
the register dump. This will help us identify which module did not
enter the idle state before executing the warm reset.

Without this patch, it is not easy for the RDs to determine the
reason for the warm reset failure.

[Release-log]
N/A


Change-Id: Ie6630d8c74aaf9234085f9c31e53ed7479606cc4