[openwrt-24][mt7988][bsp][Fix ETH driver bring up issue when eth2 SFP is enabled]
[Description]
Fix ETH driver bring up issue when eth2 SFP is enabled.

Due to the ETH2 SFP needs to use I2C1, and the pins for I2C1 are
occupied by ETH for adjusting the MDIO driving strength, once I2C1
is initialized, it causes the ETH driver to fail to initialize.
As a result, none of the ETH network interfaces are brought up.

Without this patch, the ETH driver is unable to bring up when eth2
SFP is enabled.

[Release-log]
N/A


Change-Id: I7a6cdf4c6cd81c1631ea95b5ddf39b382778d1f3