[kernel][common][hnat][Fix HNAT UNBIND issue of PPPoE WAN to LAN]
[Description]
Fix HNAT UNBIND issue of PPPoE WAN to LAN.

Without this patch, the PPPoE downlink cannot be HNAT acclerated.
Before the mac header pointer of the skb is set,
the packet destination MAC multicast check in pre-routing
might be true and block the traffic to be offloaded.

This patch also enhance the multicast detect by IP address checking
in pre-routing hook to prevent multicast to unicast traffic being
offloaded.

[Release-log]
N/A


Change-Id: I9421983ad202dbeb74463461e535b4ad06fa4ac3