[kernel-6.6][common][eth][Fix dma size and dim function in the rss patch]
[Description]
Fix dma size and dim function in the rss patch.

When we previously added the rss patch, we missed the configuration
of dma size and dim tx function, which may lead to SRAM overflow and
interrupt delay time errors.

Additionally, we found that the dim tx/rx function settings for QDMA
delay interrupt seem to be reversed. This patch also modifies the
relevant parts.

[Release-log]
N/A


Change-Id: I62242b6631eec925067205d452761c3c7d6526ab