[openwrt-24][mt7987][add and fix usb support]
[Description]
Add USB support for mt7987 to ensure proper functionality

Fix USB issue when only U2 port is enabled. It appears that when
only the U2 port is active, the infra_usb_pipe_ck_p1 clock remains
disabled, causing USB probe failures in OpenWrt 24.10 with USB2 port
enabled but U3 port disabled

To resolve this, infra_usb_pipe_ck_p1 is marked as a critical clock.
In the previous 21.02 SDK, this issue did not occur because the
driver was built-in, and the clock management flow was as follows:
- infra_usb_utmi_ck_p1 and infra_usb_pipe_ck_p1 were enabled by default.
- USB probe successfully read registers, and the probe was successful.
- Linux disabled unused clocks, keeping infra_usb_utmi_ck_p1 enabled
  while infra_usb_pipe_ck_p1 was turned off.

In contrast, in 24.10, with the USB driver as a built kernel module,
the flow changes to:
- infra_usb_utmi_ck_p1 and infra_usb_pipe_ck_p1 are enabled by default.
- Linux disables unused clocks, turning off both infra_usb_utmi_ck_p1
  and infra_usb_pipe_ck_p1.
- Upon inserting the kernel module, USB probe attempts to enable
  infra_usb_utmi_ck_p1, but fails to read registers, resulting in
  a probe failure.

[Release-log]
N/A

Change-Id: I9650eb627c2cff70272cf2d148e4120a19e61013