[openwrt-24][MAC80211][hnat][Add DSCP learning flow to the nft_flow_offload]
[Description]
Add DSCP learning flow to the nft_flow_offload.

We have implemented a mechanism to learn the DSCP until count of the
bi-directional TCP packets reached 30.

If a user would like to enable this feature, they can use the command
below to activate it.
 - echo 1 > /proc/sys/net/netfilter/nf_conntrack_qos

Without this patch, the PPEs are unable to learn the bi-directional DSCP
from SKBs.

[Release-log]
N/A


Change-Id: I3eeb47d5795fc7e0f7e84ebf0506efa76e5e7d70