[kernel][mt7988][hnat][Fix GRE/IPSec and VXLAN/IPSec decap unbind issue]
[Description]
Fix GRE/IPSec and VXLAN/IPSec decap unbind issue.
Currently, TOPS only supports L2 GRE and L2 VXLAN which means that
original packet's ethernet header will be encapsulated within the
tunnel. The MAC address of this inner eth header should not be altered.
However, in the IPSec binding flow, the MAC address of the PPE entry is
set to the "outer" MAC address.
We should add a condition that don't modify mac address when tops_entry
is set.

[Release-log]
N/A


Change-Id: I6374cafd7394698964b3a7efbec3b1f15568c57c