[HIGH][kernel][common][eth][Add mtk_stats_update_mac() to the GDM TX/RX monitors for the NETSYS SER]
[Description]
Add mtk_stats_update_mac() to the GDM TX/RX monitors for the NETSYS SER.

The GDM TX/RX monitors use the GDM counters to determine whether the
GMAC/GDM is hanging. Therefore, we need to force an update of the GDM
counters to prevent the monitors from using outdated counters, which
could lead to false alarms.

Without this patch, the user may encounter a false alarm when GMAC
connects with an external 1G/2.5G PHY.

[Release-log]
N/A


Change-Id: If12eba186de1b4f4372ef177ca881186ff632c43