[MAC80211][WiFi6][MT76][Fix incorrect MIC calculation using certain cipher]
[Description]
Fix incorrect MIC calculation using certain cipher.

If IGTK is set after GTK to FW, it seems like the length of MIC
somehow becomes incorrect.

Also, currently only IGTK with cipher CMAC-128 is set to FW, while
other ciphers are reported as not supported and MAC80211 will handle
the MIC calculation.

As discussed, we decided to report cipher CMAC-128 as not supported,
not only to solve the problem but also to maintain the consistency of
setting IGTK.

[Release-log]
N/A

Change-Id: Iad434b8faef92c8d3008dd24d30d91bd2bf03102