[kernel-6.6][mt7988][eth][Change AN8831X 10G PHY to use 10GBASE-R exclusively on Bpi-R4-Pro]
[Description]
Change AN8831X 10G PHY to use 10GBASE-R exclusively on Bpi-R4-Pro.

Force the AN8831X 10G PHY to use 10GBASE-R for connecting to the
Ethernet MAC. This removes the need to control the MaxLinear interface
for SFP in 10GBASE-R or the 10G PHY in USXGMII when an SFP module is
plugged in or unplugged.

[Release-log]
N/A


Change-Id: I33077485bd2301d3f700b39d6ba352162f214ec9