[openwrt-24][common][autobuild][add command line support for setting feeds commit ID]
[Description]
Add support for specifying the feeds commit ID from the command line.
This enhancement allows users to dynamically set the commit ID for
the feeds during the build process. Note that you can only set the
commit ID from the console when `feed_revision` does not exist in
the autobuild target folder or its sub-folder.

[Release-log]
N/A

Change-Id: I22c7e38d28b0d7ba7be596d9c5afc866124e8ceb