[kernel][mt7988/87][thermal][Change cpufreq cooling registration to use cpufreq framework]
[Description]
Change cpufreq cooling registration to use cpufreq framework.

This change updates the registration process to utilize the
cpufreq framework, improving maintainability and consistency.

We use the Linux kernel cpufreq cooling framework to dynamically
adjust the CPU frequency based on the Tj temperature. For mt7988
only, the Vproc voltage is also adjusted.

By default, the thermal framework uses the step-wise governor to
change the cooling device's cooling state. When the Tj temperature
rises above or falls below the trip point, the CPU frequency is
adjusted in steps. There is no conflict when the step-wise governor
and the ondemand governor operate simultaneously. Regardless of
whether the temperature is above or below the trip point, the
frequency state adjusted by the ondemand governor will not exceed
the current state of the cpufreq cooling device.

By default, we set the trip point for frequency scaling to 117°C
in order to prevent the IC from triggering a thermal reset.

Without this patch, there is no proactive cooling behavior before
triggering a thermal reset.

[Release-log]
N/A


Change-Id: Ied6809149f2b7e30b5c18b9dae3e5952fc95d33e