[openwrt][pcie][mt7987][Add wifi reset GPIO support for pcie0]
[Description]
Add support to assert GPIO38 during the PCIe driver probe. This
one part of the implementation for the WiFi reset feature. It may
require coordination with the OpenWrt/RDKB/Prpl scripts and reference
circulars due to differences between various WiFi chipsets.
Please refer to the WiFi reset-related AN and upgrade any
related code/scripts/circulars if needed.

[Release-log]
N/A

Change-Id: Ic533f0c8f494ca42ffd9befbf48cb64fdb0becc6