[kernel-6.6][common][eth][Fix panic issue with napi_enable in mtk_open]
[Description]
Fix panic issue with napi_enable in mtk_open.

In some scenarios, the ETH may be initialized late, making it very
close to when mtk_open is called. This could lead to a kernel panic at
netif_napi_enable because the ETH driver does not call netif_napi_add().

[Release-log]
N/A


Change-Id: Ibe2749cdd29021dfd1d4c56b51d1f8ac3c025634