[openwrt-24][MAC80211][hnat][Add L2 bridge VLAN support to the xt_FLOWOFFLOAD]
[Description]
Add L2 bridge VLAN support to the xt_FLOWOFFLOAD.

Our bridging OFFLOAD of the xt_FLOWOFFLOAD is based on the Bridge
NETFILTER module, which passes L2 traffic to the L3 forward rule.
However, it encounters an FDB lookup issue at br_fill_forward_path()
because it does not have VLAN ID information when Bridge VLAN is
activated. Therefore, we propose a method to record VLAN information
in the ctx structure to assist the L2 traffic in passing the checking
rule mentioned previously.

Without this patch, the user will experience low throughput issues in
the L2 traffic when the Bridge VLAN is activated.

[Release-log]
N/A


Change-Id: I16a121be7985b53d8ab6c86a10e50b2e609cfee7