[openwrt-24][MAC80211][hnat][Add L2 bridge VLAN support to the nft_flow_offload]
[Description]
Add L2 bridge VLAN support to the nft_flow_offload.

Our bridging OFFLOAD of the nft_flow_offload is based on the Bridge
NETFILTER module, which passes L2 traffic to the L3 forward rule.
However, it encounters an FDB lookup issue at br_fill_forward_path()
because it does not have VLAN ID information when Bridge VLAN is
activated. Therefore, we propose a method to record VLAN information
in the ctx structure to assist the L2 traffic in passing the checking
rule mentioned previously.

Without this patch, the user will experience low throughput issues in
the L2 traffic when the Bridge VLAN is activated.

[Release-log]
N/A


Change-Id: I07f733a3475793d700f2198c57835c8c533d4202