[openwrt][mt7987][usb][Fix USB issue when only U2 port is enabled]
[Description]
Fix a USB issue that occurs when only the USB2 port is enabled but
the U3 port is disabled. In this case, the `infra_usb_pipe_ck_p1`
clock remains disabled, causing USB failures.

This change ensures that the clock is properly managed, allowing for
successful USB initialization.

[Release-log]
N/A

Change-Id: Ia92908f737e16f8c8b25bbd1aac24464e32ec276